% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NbClust.R
\name{results_to_rankings}
\alias{results_to_rankings}
\title{Takes a matrix where the columns represent a value $k$ for the number of 
clusters and each of the rows is a voter, this is, a index giving its values
for each number of cluster.}
\usage{
results_to_rankings(res)
}
\arguments{
\item{res}{A matrix witht the results of each index for each partition}
}
\value{
A profile of rankings where the voters are the indexes and the 
candidates the number of clusters
}
\description{
A ranking is created for each index depending on how they would be sorted
}
